Add CRUD and seeder for booking status enum

+++ Infrastructure/Repositories/EnumMasterData/EnumMasterDataRepository.cs
@@ -3,7 +3,7 @@ using Infrastructure.Contexts;
 
 namespace Infrastructure.Repositories.EnumMasterData
 {
-    public class EnumMasterDataRepository : RepositoryAsync<Domain.Entities.EnumMasterData.EnumMasterData, long>, IEnumMasterDataRepository
+    public class EnumMasterDataRepository : RepositoryAsync<Domain.Entities.EnumMasterData.EnumMasterData, int>, IEnumMasterDataRepository
     {
         public EnumMasterDataRepository(ApplicationDbContext dbContext) : base(dbContext)
         {
